5-Eletiva-Python/secao-6-Django-II/dia-1-Autenticação/personal-trainer-exercise/dev-requirements.txt
argh==0.31.3
asgiref==3.8.1
Django==5.0.7
django-static-autocollect==0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
pathtools==0.1.2
pillow==10.4.0
PyJWT==2.8.0
PyYAML==6.0.1
sqlparse==0.5.0
watchdog==0.8.3
whitenoise==6.7.0
